+++ docs/ReleaseNotes.md
@@ -1,6 +1,9 @@
 // Before any commit, use the command line: tsc --noEmit
 
-# 0.1.5.7 Alert popUps refactor
+# 0.1.6.0 Document export module
 
-- Added font color for text without color defined.
-- feedbacks callback are now the last callback in the function definition.
+- Added Media folder to projectFolder creation
+- docx dependency installed
+- expo-sharing dependency installed
+- New FileExportModules folder. Folder responsible to isolate the logic of file mounting of many File formats.
+- New File Export Service. Responsable to low level access to app files to store other files like Docx, Pdf, csv, etc.